// src/app/modules/payment/payment.interface.ts
import { Types } from "mongoose";

export enum PAYMENT_STATUS {
  PAID = "PAID",
  UNPAID = "UNPAID",
  REFUNDED = "REFUNDED",
  CANCELLED = "CANCELLED",
  FAILED = "FAILED",
}

export interface IPayment {
  booking: Types.ObjectId;
  transactionId: string;
  amount: number;
  // eslint-disable-next-line @typescript-eslint/no-explicit-any
  paymentGatewayData?: any;
  invoiceUrl?: string;
  status : PAYMENT_STATUS
}